feat(service): Add metricas

+++ src/main/java/com/example/SalesHub/dto/filter/EstoqueFilter.java
@@ -1,10 +1,11 @@
 package com.example.SalesHub.dto.filter;
 
 import lombok.Builder;
+import java.math.BigDecimal;
 
 @Builder
 public record EstoqueFilter(
         Long id,
         Long produtoId,
-        Long quantidadeInicial,
-        Long quantidadeAtual) { }
+        BigDecimal quantidadeInicial,
+        BigDecimal quantidadeAtual) { }
