+++ src/main/java/com/example/SalesHub/dto/filter/HistoricoFilter.java
@@ -3,6 +3,7 @@ package com.example.SalesHub.dto.filter;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Builder
@@ -11,7 +12,7 @@ public record HistoricoFilter(
         Long usuarioId,
         Long produtoId,
         Long estoqueId,
-        Long quantidadeRetirada,
+        BigDecimal quantidadeRetirada,
         @JsonFormat(pattern = "dd/MM/yyyy")
         LocalDate dataInicial,
         @JsonFormat(pattern = "dd/MM/yyyy")
