+++ src/main/java/com/example/SalesHub/dto/filter/VendaFilter.java
@@ -2,7 +2,6 @@ package com.example.SalesHub.dto.filter;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Builder;
-
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
@@ -10,8 +9,8 @@ import java.time.LocalDate;
 public record VendaFilter(
         Long id,
         Long usuarioId,
-        Long valor,
-        Long quantidade,
+        BigDecimal valor,
+        BigDecimal quantidade,
         BigDecimal valorTotalVendas,
         @JsonFormat(pattern = "dd/MM/yyyy")
         LocalDate dataInicial,
