+++ src/main/java/com/example/SalesHub/dto/projection/EstoqueProjection.java
@@ -1,10 +1,11 @@
 package com.example.SalesHub.dto.projection;
 
 import lombok.Builder;
+import java.math.BigDecimal;
 
 @Builder
 public record EstoqueProjection(
         Long id,
         Long produtoId,
-        Long quantidadeInicial,
-        Long quantidadeAtual) {}
+        BigDecimal quantidadeInicial,
+        BigDecimal quantidadeAtual) {}
